Make EVM emulator tests common
